requirements.txt
streamlit
langchain
bs4
requests
python-dotenv
werkzeug
langchain_community
